backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27
python-dotenv==1.0.0
pydantic==2.4.2
websockets==12.0
aiohttp==3.9.1
python-magic-bin==0.4.14; sys_platform == 'win32'
numpy>=1.20.0,<1.22.0
opencv-python>=4.5.0,<4.6.0
Pillow>=8.0.0,<9.0.0
aiofiles
psutil==5.9.8
prometheus-client==0.19.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
radon==6.0.1
coverage==7.4.1
pytest==7.4.4
pytest-cov==4.1.0